v2018.03.22: Fix CDRVR crashing when TPA segment for page 0 is uninitialized.
CDRVR was calling the driver routine by invoking CALSLT in the TPA
segment switched in page 0. However, if the system has booted in BASIC
this segment is not initialized and there is nothing at the CALSLT
entry point; thus the system crashed. This could be easily reproduced
by using FDISK.