Fix the update of UD_ACLU in FR_CHAIN.
FR_CHAIN was simply updating UD_ACLU to the first cluster number
in the freed chain, but this was wrongly assuming that cluster numbers
in a FAT chain always increase, while they could actually decrease
(e.g. the chain could be 200-201-202-100-FFF).

The fix implemented in this commit uses FAT_CHAIN, a variable used only
by AL_CLUSTERS, to temporarliy keep track of which is the lowest
cluster number freed during the process, then uses this value instead
of the first cluster number in the chain to update UD_ACLU
if appropriate.