Move UD_ACLU to the end of the unit descriptor to prevent problems
The improvement of the cluster chain allocation process performance
introduced a new unit descriptor field, UD_ACLU, in the middle
of the unit descriptor (between UD_MBYTE and UD_CDIR). This caused
severe problems when writing data (weird errors, computer crashes,
even data corruption) because somewhere in the kernel there's code
that assumed that these fields are stored consecutively in memory.

The fix consists of moving UD_ACLU to the end of the unit descriptor.