Cleanup kernel code to remove warnings and errors at compile time.
The kernel code was throwing some warnings and errors on compile,
like unknown and duplicate labels, which didn't prevent the kernel
ROMS to be properly generated bute were anyway quite confusing and
masked possible real errors. This commit does the appropriate changes
to code files so that these warnings and errors disappear.