Remove references to the NEXTOR.DAT file
Old versions of the Nextor kernel searched for a NEXTOR.DAT file
as part of the automatic drive mapping procedure, but in v2.1.0 beta 2
this was changed to checking the "active" flag in the partition table
entry instead. Some parts of the documentation were still referencing
that NEXTOR.DAT file, these have been corrected.