Fix: incorrect setting of UD_ACLU when freeing a FAT chain
When freeing a FAT chain (while deleting a file) the UD_ACLU value
of the unit descriptor, which holds the smaller free cluster number,
is updated to the smallest cluster number freed. The code doing this
had a bug and instead, a random memory address was modified; for
small cluster numbers this meant that the Nextor code itself was
overwritten and this caused the computer to crash. This commit
fixes that.