Add a new KIILLDSKIO environment variable.
When this variable is set to "ON" (case insensitive) the 512 byte buffer
that is normally allocated for the DSKI$ and DSKO$ BASIC commands won't
be allocated when starting the BASIC environment, thus these 512 bytes
will be gained back for BASIC programs. This implies that the DSKI$ and
DSKO$ commands will be disabled (they will throw "syntax error").