Modify ZAP_ALL to skip drives assigned to device-based drivers.
ZAP_ALL deletes all unit descriptors, and is called
when _FLUSH with the "invalidate" option is executed.
While on MSX-DOS and drive-based drivers this has no
negative consequences, for units on device-based drivers
this implies that the drive is reverted to its default
state, which may mean "unmapped". Thus these must be
skipped in the zapping process.